More API generator fixes, drastically simplified user model
