+++ generator/api_model/templates/partials/resource-collection-options.js
@@ -1,0 +1,4 @@
+const collection_options = {
+  timestamps: true,
+  versionKey: false
+}
